refactor(server): emit integral JSON for long & counting progress
The long and counting progress emitters were widening their values to double,
so a countdown reported "progress": 1.0 … 5.0 on the wire. Those are counts —
emit them as whole-number JSON (1 … 5) instead. The double and percentage
emitters keep floating-point output (percent stays consistent, e.g. complete(1.0)
sends 1.0). The wire numeric type now follows the emitter type, not the value.

Both forms are valid per the spec's `number` type (which MAY be floating point);
integers are also friendlier to strict-integer clients. ProgressChannel now
carries the value as a Number and serializes integral vs floating accordingly;
the strictly-increasing check still compares in double.

+++ mocapi-server/src/main/java/com/callibrity/mocapi/server/progress/DefaultCountingProgressEmitter.java
@@ -33,7 +33,7 @@ class DefaultCountingProgressEmitter implements CountingProgressEmitter {
 
   @Override
   public void emit(String message) {
-    channel.emit((double) ++count, message);
+    channel.emit(++count, message);
   }
 
   @Override
